+++ lib/massive_record/orm/schema/field.rb
@@ -13,9 +13,9 @@ module MassiveRecord
           :hash => {},
           :date => lambda { Date.today },
           :time => lambda { Time.now }
-        }
+        }.freeze
 
-        TYPES = TYPES_DEFAULTS_TO.keys
+        TYPES = TYPES_DEFAULTS_TO.keys.freeze
 
         attr_writer :default
         attr_accessor :name, :column, :type, :fields, :coder, :allow_nil
@@ -109,11 +109,13 @@ module MassiveRecord
 
 
         def decode(value)
+          value = value.force_encoding(Encoding::UTF_8) if utf_8_encoded? && !value.frozen? && value.respond_to?(:force_encoding) 
+
           return value if value.nil? || value_is_already_decoded?(value)
           
           value = case type
                   when :boolean
-                    value.blank? ? nil : !value.to_s.match(/^(true|1)$/i).nil?
+                    value.blank? || value == @@encoded_nil_value ? nil : !value.to_s.match(/^(true|1)$/i).nil?
                   when :date
                     value.blank? || value.to_s == "0" ? nil : (Date.parse(value) rescue nil)
                   when :time
@@ -123,7 +125,13 @@ module MassiveRecord
                       value = value.to_s if value.is_a? Symbol
                       coder.load(value)
                     end
-                  when :integer, :float, :array, :hash
+                  when :integer
+                    if value =~ /\A\d*\Z/
+                      coder.load(value) if value.present?
+                    else
+                      hex_string_to_integer(value)
+                    end
+                  when :float, :array, :hash
                     coder.load(value) if value.present?
                   else
                     raise "Unable to decode #{value}, class: #{value}"
@@ -135,11 +143,11 @@ module MassiveRecord
         end
 
         def encode(value)
-          if type == :string && !(value.nil? || value == @@encoded_nil_value)
+          if value.nil? || should_not_be_encoded?
             value
           else
             value = value.try(:utc) if Base.time_zone_aware_attributes && field_affected_by_time_zone_awareness?
-            coder.dump(value)
+            coder.dump(value).to_s
           end
         end
 
@@ -183,6 +191,18 @@ module MassiveRecord
 
         def field_affected_by_time_zone_awareness?
           type == :time
+        end
+
+        def hex_string_to_integer(string)
+          Wrapper::Cell.hex_string_to_integer(string)
+        end
+
+        def utf_8_encoded?
+          type != :integer
+        end
+
+        def should_not_be_encoded?
+          [:string, :integer].include?(type)
         end
       end
     end